Delete in push function now check version to avoid conflicts and docs for push are finished

+++ serverApp/FileSynchroniser/src/routes/files.js
@@ -23,7 +23,6 @@ router.post('/push', (req, res)=> {
         if (req.files.file) {
             let userFile = req.files.file;            
             //Look in DB to see if file already exists
-            //TODO: check if it exists on the file system (DB and FS are out of sync)
             File.findOne({ filename: userFile.name }, (err, file)=>{
                 if (err) {
                     console.error(err);
@@ -47,7 +46,7 @@ router.post('/push', (req, res)=> {
                                 fs.mkdirSync(global.fileDirectory, { recursive: true });
                             }
                             //Move file to the directory
-                            userFile.mv(path.join(global.fileDirectory, userFile.name), (err)=> {
+                            userFile.mv(path.join(file.path), (err)=> {
                                 if (err) {
                                     console.error(err);
                                     return res.status(500).send({success: false, error:'Error saving the file'});
@@ -114,7 +113,6 @@ router.post('/push', (req, res)=> {
         //Check if delete flag is true
         if( req.body.delete === "true" ){ //delete == true
             //Look in the DB to see if file exists
-            //TODO: check if it exists on the file system (DB and FS are out of sync)
             File.findOne({ filename: req.body.filename }, (err, file)=>{
                 if (err) {
                     console.error(err);
@@ -122,23 +120,28 @@ router.post('/push', (req, res)=> {
                 }
                 //Check that file exists
                 if(file){
-                    //Delete/unlink the file
-                    fs.unlink(file.path, (err) => {
-                        if (err) {
-                            console.error(err);
-                            return res.status(500).send({success: false, error: "Error deleting file"});
-                        }
-                        //Delete file info from DB
-                        File.deleteMany({ filename: req.body.filename }, function (err) {
+                    //Check if file is last version
+                    if(parseInt(req.body.version) === file.version){ //Files versions are the same so there is no conflict
+                        //Delete/unlink the file
+                        fs.unlink(file.path, (err) => {
                             if (err) {
-                                //TODO: should resave the file to the dir
                                 console.error(err);
-                                return res.status(500).send({success: false, error: "Error deleting file info from database"});
+                                return res.status(500).send({success: false, error: "Error deleting file"});
                             }
-                            //File deleted successfully
-                            return res.send({success: true, file: null});
+                            //Delete file info from DB
+                            File.deleteMany({ filename: req.body.filename }, function (err) {
+                                if (err) {
+                                    //TODO: should resave the file to the dir
+                                    console.error(err);
+                                    return res.status(500).send({success: false, error: "Error deleting file info from database"});
+                                }
+                                //File deleted successfully
+                                return res.send({success: true, file: null});
+                            });
                         });
-                    });
+                    }else{ //File versions are different HTTP 409 == Conflict
+                        res.status(409).send({success: false, error: "The file was not up to date, not deleting it to avoid conflicts"})
+                    }
                 }else{ //file does not exist send error
                     return res.status(400).send({success: false, error:'File to delete does not exist'});
                 }
